requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.2
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
Brlapi==0.8.3
cachetools==5.3.0
celery==5.3.1
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
cmake==3.26.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.8
contourpy==1.1.0
cryptography==41.0.1
cupshelpers==1.0
cycler==0.11.0
databricks-cli==0.17.7
dataclasses-json==0.5.9
dbus-python==1.2.18
defer==1.0.6
dicomweb-client==0.59.1
dictdiffer==0.9.0
diskcache==5.6.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.2
docarray==0.21.0
docker==6.1.3
docx2txt==0.8
dpath==2.1.6
dulwich==0.21.5
duplicity==0.8.21
dvc==3.4.0
dvc-data==2.3.3
dvc-http==2.30.2
dvc-objects==0.23.0
dvc-render==0.5.3
dvc-studio-client==0.11.0
dvc-task==0.3.0
dvclive==2.12.0
ecdsa==0.18.0
einops==0.6.1
emoji==2.6.0
entrypoints==0.4
exceptiongroup==1.1.1
expiring-dict==1.1.0
expiringdict==1.2.2
fastapi==0.95.0
fasteners==0.14.1
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
funcy==2.0
future==0.18.2
gdown==4.7.1
girder-client==3.1.17
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
gptcache==0.1.35
grandalf==0.8
greenlet==2.0.2
grpcio==1.56.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
hydra-core==1.3.2
idna==3.3
imageio==2.31.1
importlib-metadata==4.6.4
iterative-telemetry==0.0.8
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.1.2
jeepney==0.7.1
jina-hubble-sdk==0.39.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.3.1
langchain==0.0.141
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lit==16.0.6
lmdb==1.4.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
mlflow==2.4.1
monai==1.2.0
monailabel==0.7.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nanotime==0.5.2
netifaces==0.11.0
networkx==3.1
nibabel==5.1.0
ninja==1.11.1
nltk==3.8.1
nptyping==2.5.0
numpy==1.25.0
numpymaxflow==0.0.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python-headless==4.7.0.72
openslide-python==1.1.2
orjson==3.9.1
packaging==23.1
pandas==2.0.3
paramiko==2.9.3
passlib==1.7.4
pathspec==0.11.1
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.10
pydicom==2.3.1
pydicom-seg==0.4.1
pydot==1.4.2
pygit2==1.12.2
Pygments==2.15.1
PyGObject==3.42.1
pygtrie==2.5.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynetdicom==2.0.2
pynrrd==1.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
pyspark==3.4.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytorch-ignite==0.4.11
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0
querystring-parser==1.2.4
regex==2023.6.3
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.4
rfc3986==1.5.0
rich==13.4.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
schedule==1.1.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
scmrepo==1.0.4
SecretStorage==3.3.1
shapely==2.0.1
shortuuid==1.0.11
shtab==1.6.2
SimpleITK==2.2.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
sqltrie==0.7.0
stanza==1.5.0
starlette==0.26.1
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thinkgpt @ git+https://github.com/alaeddine-13/thinkgpt.git@59dd7e8a881d7c6cf280c140344a72321bc289d6
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.3.3
timeloop==1.0.2
tomlkit==0.11.8
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
tradingview-ta==3.3.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.16
usb-creator==0.3.7
uvicorn==0.21.1
vine==5.0.0
virtualenv==20.23.1
voluptuous==0.13.1
wadllib==1.3.6
Wand==0.6.11
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.3.6
xdg==5
xkit==0.0.0
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==1.0.0
